Shader FieldInfo models across D2D1/D3D12 generators
